requirements.txt
websockets==12.0
fastapi==0.110.0
pydantic-settings==2.2.1
uvicorn==0.29.0
aioredis==2.0.1